+++ src/ParquetMain.py
@@ -12,9 +12,9 @@ def main(cMgr):
     logging.basicConfig(level=logging.INFO, format=au.LOG_FORMAT)
     # workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['parquet', 'summarize'])
     # workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['summarize', 'simSummary'])
-    # workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['simSummary'])
+    workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['simSummary'])
     # workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['parquet', 'summarize', 'simSummary'])
-    workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['createPDFCache'])
+    # workitemQueues = sm.generateWorkitems(cMgr, phasesToInclude=['createPDFCache'])
     sm.main(cMgr, workitemQueues=workitemQueues)